+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-16T09:29:00.126Z",
+    "2026-06-16T10:59:02.870Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {