+++ packages/orq-rc/RELEASES.md
@@ -10029,3 +10029,13 @@ Based on:
 - [typescript v4.11.0-rc.24] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.24 - packages/orq-rc
+
+## 2026-06-16 11:01:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.778.0 (2.904.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.25] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.25] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.25 - packages/orq-rc